Extract ModelResponseStreamEventType
